fix(translation): remove shared mutable pipeline and test deps
Pass glossary/TM match resolution only through search input instead of
mutating pipeline instances, and move CAT recommendation orchestration to
cat.ts so tests can mock dependencies with vi.mock.
