+++ apps/hyperlocalise-web/src/lib/translation/cat.ts
@@ -27,7 +27,17 @@ import {
   GlossaryConcordanceService,
   TranslationMemoryConcordanceService,
 } from "@/lib/translation/concordance";
-import type { CatAiRecommendationInput } from "@/lib/translation/domain";
+import { assembleStringTranslationContextSnapshot } from "@/lib/translation/context";
+import type {
+  CatAiRecommendationError,
+  CatAiRecommendationInput,
+  CatAiRecommendationResult,
+} from "@/lib/translation/domain";
+import {
+  CatRecommendationEngine,
+  loadOrganizationTranslationModel,
+} from "@/lib/translation/generation";
+import { err, ok, type Result } from "@/lib/primitives/result/results";
 
 export type {
   CatVisualContext,
@@ -413,4 +423,64 @@ export function mapCatConcordanceForAiRecommendation(
       targetLocale,
     })),
   };
+}
+
+export async function generateCatAiRecommendation(
+  input: CatAiRecommendationInput,
+): Promise<Result<CatAiRecommendationResult, CatAiRecommendationError>> {
+  const modelResult = await loadOrganizationTranslationModel(input.projectId);
+  if (!modelResult.ok) {
+    return err({ code: modelResult.code, message: modelResult.message });
+  }
+
+  const hasPreloadedConcordance =
+    input.glossaryTerms !== undefined || input.translationMemoryMatches !== undefined;
+
+  const contextResult = await assembleStringTranslationContextSnapshot(
+    input.projectId,
+    {
+      sourceLocale: input.sourceLocale,
+      targetLocales: [input.targetLocale],
+      sourceText: input.sourceText,
+      context: input.context ?? undefined,
+      maxLength: input.maxLength,
+      metadata: {
+        sourcePath: input.sourcePath,
+        key: input.key,
+      },
+    },
+    undefined,
+    {
+      organizationId: input.organizationId,
+      glossaryMatchResolution: defaultGlossaryMatchResolution,
+      translationMemoryMatchResolution: defaultTranslationMemoryMatchResolution,
+      skipConcordance: hasPreloadedConcordance,
+    },
+  );
+
+  if (!contextResult.ok) {
+    return err({
+      code: "translation_context_assembly_failed",
+      message: contextResult.message,
+    });
+  }
+
+  try {
+    const recommendation = await new CatRecommendationEngine(modelResult.model).recommend(input, {
+      projectName: contextResult.snapshot.project.name,
+      projectTranslationContext: contextResult.snapshot.project.translationContext,
+      knowledgeMemory: contextResult.snapshot.knowledgeMemory ?? undefined,
+      glossaryTerms: input.glossaryTerms ?? contextResult.snapshot.glossaryTerms ?? [],
+      translationMemoryMatches:
+        input.translationMemoryMatches ?? contextResult.snapshot.translationMemoryMatches ?? [],
+    });
+
+    return ok(recommendation);
+  } catch (error) {
+    return err({
+      code: "ai_recommendation_failed",
+      message:
+        error instanceof Error ? error.message : "Failed to generate AI translation recommendation",
+    });
+  }
 }
